ETL Breast Cancer Report - Co #6
Breast Cancer has affected the lives of many individuals across the United States. Subsequently, many questions have arisen regarding the cause of such a caustic and life-threatening disease. With so many questions from the public, our client asked us to focus on the correlation between income level and the rate of breast cancer cases being discovered per county in the eastern portion of the United States.  
The expected outcome from the ETL project is one supporting a negative correlation between rising incomes and an increasing rate of cancer cases per county in our selected data points, the majority of which are located in the Eastern portion of the United States.